Add files via upload
V1.2  WIP.  Dogs Trust mostly working, but need to add functionality to move to next page, which means probably pulling out the code that loads a list page and searches for a dog string to build a list of URLs.  SNDogs and DogsTrust v. similar for that.  Similarly goToNextDog could probably be pulled back into the base class, because now it just goes through a pre-built list.